+++ tests/codeception/_support/AcceptanceTester.php
@@ -1,6 +1,8 @@
 <?php
 namespace gallery;
 
+use Codeception\Util\Locator;
+
 /**
  * Inherited Methods
  * @method void wantToTest($text)
@@ -20,15 +22,118 @@ class AcceptanceTester extends \AcceptanceTester
 {
     use _generated\AcceptanceTesterActions;
 
-   /**
-    * Define custom actions here
-    */
-    private function enableSpaceModule(AcceptanceTester $I)
+    public function enableModule($guid, $moduleId)
     {
-        $I->amAdmin();
-        $I->wantToTest('the creation of a task list');
-        $I->amGoingTo('install the calendar module for space 1');
-        $I->enableModule(1, 'gallery');
-        $I->amOnSpace1();
+        $this->amOnSpace($guid, '/space/manage/module');
+        $this->seeElement('.enable-module-'.$moduleId);
+        $this->jsClick('.enable-module-'.$moduleId);
+        $this->waitForElement('.disable-module-'.$moduleId);
+        $this->amOnSpace($guid);
+    }
+
+    public function enableSpaceModule()
+    {
+        $this->amAdmin();
+        $this->wantToTest('the creation of a task list');
+        $this->amGoingTo('install the calendar module for space 1');
+        $this->enableModule(1, 'gallery');
+        $this->amUser2(true);
+        $this->amOnSpace1();
+    }
+
+    public function seeGalleryInCotnainerNav()
+    {
+        $this->expectTo('see gallery entry in the nav');
+        $this->waitForText('Gallery', null, '.layout-nav-container');
+    }
+
+    public function accessGallery()
+    {
+        $this->click('Gallery', '.layout-nav-container');
+        $this->waitForText('List of galleries');
+        $this->see('Posted Media Files');
+    }
+
+    public function createGallery($title = 'Test gallery', $description = 'My test gallery', $public = true)
+    {
+        $this->click('Click here to add new Gallery');
+        $this->waitForText('Add new gallery', null, '#globalModal');
+        $this->fillField('#customgallery-title', $title);
+        $this->fillField('#customgallery-description', $description);
+
+        if ($public) {
+            $this->click('[for="galleryeditform-visibility"]', '#globalModal');
+        }
+
+        $this->click('Save', '#globalModal');
+        $this->waitForText('Gallery '.$title, null, '#gallery-container .panel-heading');
+
+        if ($public) {
+            $this->see('Public');
+        } else {
+            $this->dontSee('Public');
+        }
+    }
+
+    public function uploadMedia($file = 'logo.jpg', $shouldFail = false)
+    {
+        $this->attachFile('#gallery-media-upload', $file);
+        if (!$shouldFail) {
+            $this->waitForElementVisible(Locator::elementAt('.gallery-list-entry', 2));
+        }
+    }
+
+    public function editMedia()
+    {
+        $this->clickGalleryItemDropDown('Edit');
+        $this->waitForText('Edit media', null, '#globalModal');
+        $this->fillField('#media-description', 'My new media!');
+        $this->click('Save', '#globalModal');
+        $this->seeSuccess('Saved');
+    }
+
+    public function seeMediaInStream()
+    {
+        $this->click('Stream', '.layout-nav-container');
+        $this->waitForText('My new media!', null, '.wall-entry');
+        $this->click('Open Gallery', '.wall-entry');
+        $this->waitForElement('#gallery-container .panel-heading');
+        $this->waitForText('Gallery Test gallery');
+    }
+
+    public function dontseeMediaInStream()
+    {
+        $this->click('Stream', '.layout-nav-container');
+        $this->waitForElementVisible('#wallStream .wall-entry');
+        $this->dontSee('My new media!', null, '.wall-entry');
+    }
+
+    public function deleteGalleryItem()
+    {
+        $this->clickGalleryItemDropDown('Delete');
+        $this->waitForText('Confirm delete item', null, '#globalModalConfirm');
+        $this->click('Confirm', '#globalModalConfirm');
+    }
+
+    public function clickGalleryItemDropDown($item)
+    {
+        $this->wait(2);
+        $this->click('.gallery-list-entry .dropdown-toggle');
+        $this->wait(1);
+        $this->click($item, '.gallery-list-entry .dropdown');
+    }
+
+    public function deleteGallery()
+    {
+        $this->waitForElementVisible(Locator::elementAt('.gallery-list-entry', 3));
+        $this->deleteGalleryItem();
+        $this->waitForElementNotVisible(Locator::elementAt('.gallery-list-entry', 3));
+    }
+
+    public function deleteMedia()
+    {
+        $this->waitForElementVisible(Locator::elementAt('.gallery-list-entry', 2));
+        $this->deleteGalleryItem();
+        $this->waitForElementNotVisible(Locator::elementAt('.gallery-list-entry', 2));
     }
 }